refactore Code, rearange methods
