+++ game/src/ecs/components/ai/AIComponent.java
@@ -38,7 +38,7 @@ public class AIComponent extends Component {
     public AIComponent(@DSLContextMember(name = "entity") Entity entity) {
         super(entity, name);
         System.out.println("DEBUG AI");
-        idleAI = new RadiusWalk(5);
+        idleAI = new RadiusWalk(5, 2);
         transitionAI = new RangeTransition(1.5f);
         fightAI =
                 entity1 -> {
