--- game/test/ecs/ai/MeleeAITest.java
@@ -1,3 +1,0 @@
-package ecs.ai;
-
-public class MeleeAITest {}
