--- game/test/ecs/ai/RadiusWalkTest.java
@@ -1,3 +1,0 @@
-package ecs.ai;
-
-public class RadiusWalkTest {}
