--- game/test/ecs/ai/RangeTransitionTest.java
@@ -1,3 +1,0 @@
-package ecs.ai;
-
-public class RangeTransitionTest {}
